==> src/1_detect_ASB_spike_and_slab.ipynb <==
import torch
import sys
import pyro

import seaborn as sns
import matplotlib.pyplot as plt
import scipy.stats
import pandas as pd

import numpy as np

from pooledQTL import deconvolve, beta_model, gaussian_model, io_utils, asb_data, spike_and_slab

from importlib import reload

from pathlib import Path

fdr_threshold = 0.05
device = "cuda:0" if torch.cuda.is_available() else "cpu"
use_structured_guide = True # performs much better

results_dir = Path("/gpfs/commons/home/daknowles/pooledRBPs/results/RBFOX2")
results_dir.mkdir(parents=True, exist_ok=True)

reload(spike_and_slab)

dat = pd.read_csv("../results/HNRNPA1/beta_rep0.tsv.gz", sep = "\t", index_col = False)
device = "cuda:0" if torch.cuda.is_available() else "cpu"
#device = "cpu"
#data = asb_data.ASBdata.from_pandas(dat.rename(columns = {"altCount_input" : "altCount", "totalCount_input" : "totalCount" }), device = device)

learn_likelihood_concs = True
#learn_prior_prob = True
structured_model = True

prior_prob_ase = {}
prior_prob_asb = {}

for learn_likelihood_concs in [True, False]:
    for structured_model in [True, False]:
        post_stats, post_mean = spike_and_slab.fit(
            dat, 
            learn_likelihood_concs = learn_likelihood_concs, 
            structured_model = structured_model, 
            plot = False,
            device = device
        )
        prior_prob_ase[(learn_likelihood_concs,structured_model)] = post_stats['prior_prob_ase']['mean'].cpu().item()
        prior_prob_asb[(learn_likelihood_concs,structured_model)] = post_stats['prior_prob_asb']['mean'].cpu().item()

post_stats, post_mean = spike_and_slab.fit(
    dat, 
    learn_likelihood_concs = False, 
    structured_model = True, 
    plot = True,
    device = device
)

asb_idx = post_mean['assignment_asb'].argmax()

post_mean['assignment_ase'][asb_idx]
post_mean['assignment_asb'][asb_idx]
dat.iloc[asb_idx,:]

post_stats['ase']['mean'][0,asb_idx]
post_stats['asb']['mean'][0,asb_idx]
post_stats['asb_when_ase']['mean'][0,asb_idx] # asb_when_ase

dat_sub = dat[post_mean['assignment_asb'] < 0.5]
plt.scatter(dat_sub.allelic_ratio, dat_sub.altCount_IP / dat_sub.totalCount_IP)
dat_sub = dat[post_mean['assignment_asb'] > 0.5]
plt.scatter(dat_sub.allelic_ratio, dat_sub.altCount_IP / dat_sub.totalCount_IP)


prior_prob_ase

prior_prob_asb
